refactor: rename internal package sqsnotify2 to sqsnotify
Drop the redundant "2" suffix since the module path is already
github.com/koron/sqs-notify/v2.
