+++ internal/cache/cache.go
@@ -12,7 +12,7 @@ import (
 	"github.com/koron/sqs-notify/v2/internal/stage"
 )
 
-const minCapacity = 10 // sqsnotify2.maxMsg
+const minCapacity = 10 // sqsnotify.maxMsg
 
 var (
 	errCacheFound    = errors.New("cache found")
