+++ internal/sqsnotify/sqsnotify_test.go
@@ -1,4 +1,4 @@
-package sqsnotify2
+package sqsnotify
 
 import (
 	"context"
