+++ internal/sqsnotify/version.go
@@ -1,4 +1,4 @@
-package sqsnotify2
+package sqsnotify
 
 // Version is a string for sqs-notify2's version
 const Version = "v2.2.2"
